backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
openai==0.27.0
motor==3.1.1
pydantic==1.10.8
